Seprate Logs and goes to sleep if Network Issue

+++ upload.py
@@ -6,13 +6,20 @@ import sys
 import requests
 import json
 import math
+import time
+import datetime
+from pytz import timezone
 
 S_API_KEY = "3773bd8cf9594ca7a2a6c0074f73ace7"
 D_Loc = "A-Home"
 D_NAME = "Node-AH"
-BASEPATH = [Path of CSV Folder]
+BASEPATH = [Path of CSV files]
 URL_S = "http://sensoract.iiitd.edu.in:9000/upload/wavesegment"
+now = datetime.datetime.now(timezone('Asia/Kolkata'))
 while(True):
+	loop_lock = 0
+	Log_Filename = "Log_"+str(now.day) + "_" + str(now.month) + "_" + str(now.hour) + "_" + str(now.minute) + ".txt"
+	Log_File = BASEPATH + Log_Filename
 	list_of_files = glob.glob(BASEPATH+str("*.csv"))
 	print list_of_files
 	for f in list_of_files:
@@ -31,23 +38,38 @@ while(True):
 					payload3 = { 'secretkey' : S_API_KEY, "data" : { "loc" : D_Loc, "dname" : str(D_NAME), "sname" :"LightSensor", "sid" : "1", "timestamp" : timestamp, "channels" : [ { "cname" : "channel1", "unit" : "none","readings" : [light] } ] } }
 					try:
 						r1 = requests.post(URL_S, data=json.dumps(payload1), headers=headers)
+						time.sleep(1)
 						r2 = requests.post(URL_S, data=json.dumps(payload2), headers=headers)
+						time.sleep(1)
 						r3 = requests.post(URL_S, data=json.dumps(payload3), headers=headers)
+#						lo=open(Log_File,"a")
+#						lo.write("\n"+str(time.time())+"  "+"Values Uploaded: -"+str(f))
+#						lo.close()
+		
 #						lo=open(BASEPATH+"LOG_UPLOAD.txt","a")
 #						lo.write("\n"+str(time.time())+":  "+"Values Uploaded "+str(f))
 #						lo.close()
 						
 					except Exception as f2:
-						lo=open(BASEPATH+"LOG_UPLOAD.txt","a")
+						lo=open(Log_File,"a")
 						lo.write("\n"+str(time.time())+"  "+"filename: "+f +"  Error: "+f2.__str__())
 						lo.close()
-				lo=open(BASEPATH+"LOG_UPLOAD.txt","a")
-				lo.write("\n"+str(time.time())+"  "+"Upload Success. \n Now removing file: "+str(f))
-				lo.close()
-				os.remove(f)
-	
+						loop_lock = 1
+						break
+					time.sleep(2)
 		else:
-			lo=open(BASEPATH+"LOG_UPLOAD.txt","a")
+			lo=open(Log_File,"a")
 			lo.write("\n"+str(time.time())+": File in Use \n"+f)
 			lo.close()
+		time.sleep(1)
+		if loop_lock == 0:
+			lo=open(Log_File,"a")
+			lo.write("\n"+str(time.time())+"  "+"Upload Success. \n Now removing file: "+str(f))
+			lo.close()
+			os.remove(f)
+		else:
+			lo=open(Log_File,"a")
+			lo.write("\n"+str(time.time())+"  "+"Connectivity Issue, Going to sleep ")
+			lo.close()
+			
 	time.sleep(900)
